Added rule "alpha_numeric_punct"
Fails if field contains anything characters other than alphanumeric, space and this set of punctuation   ~ ! # $ % & * - _ + = | : .  (that last, the period, is in the set and not just the end of this sentence.).

Removed //--------------- from between functions as the docblocks visually set them apart just fine.

Also fixed spelling of alphanumeric. It is not a hyphenated compound word. Affected several files.

Added rule description to docs.